rename trimesh module to triangularmesh
