+++ magpylib/_src/fields/field_BH_triangularmesh.py
